remove sports field from previous pull request and prevent courses from having no sport

+++ src/services/courseSportService.ts
@@ -1,3 +1,4 @@
+import courseRepository from "../repositories/courseRepository";
 import courseSportRepository from "../repositories/courseSportRepository";
 import { genericServRepo } from "../utils/error";
 
@@ -23,6 +24,9 @@ class CourseSportService {
 
     async delete(id: number) {
         return await genericServRepo('courseSportService.delete', 'Error deleting course\'s sport', [id], async (id) => {
+            const courseSport = await courseSportRepository.getById(id);
+            const course = await courseRepository.getById(courseSport.courseId);
+            if (course.Sports.length <= 1) throw new Error('CODE400: you cannot remove the only sport remaining in a course');
             return await courseSportRepository.delete(id);
         });
     }
